% Generated by roxygen2 (4.0.1): do not edit by hand
\name{JKpp}
\alias{JKpp}
\alias{JKpp.fourpl}
\alias{JKpp.gpcm}
\alias{JKpp.gpcm4pl}
\title{Run a jackknife}
\usage{
JKpp(estobj, ...)

\method{JKpp}{fourpl}(estobj, cmeth = "mean", maxsteps = 500,
  exac = 0.001, ctrl = list())

\method{JKpp}{gpcm}(estobj, cmeth = "mean", maxsteps = 500, exac = 0.001,
  ctrl = list())

\method{JKpp}{gpcm4pl}(estobj, cmeth = "mean", maxsteps = 500,
  exac = 0.001, ctrl = list())
}
\arguments{
\item{estobj}{An object which contains a conventional Person Parameter estimation}

\item{...}{Three Points}

\item{cmeth}{Centering method, to summarize the jackknife results to one single ability estimate}

\item{maxsteps}{The maximum number of steps the NR Algorithm will take.}

\item{exac}{How accurate are the estimates supposed to be? Default is 0.001.}

\item{ctrl}{more controls}
}
\description{
This function uses a jackknife approach to compute person parameters.
}

