% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRd.R
\name{pp_amt}
\alias{pp_amt}
\title{Real data set}
\format{A data.frame with 60 rows and 14 columns.}
\description{
This dataset contains real data from the 'Adaptive Matrices Test' (AMT).
Assesses logical reasoning as an indicator of general intelligence. The ability to identify regularities and draw logical conclusions is a very good predictor of long-term success at work.
}
\details{
The data are provided from the Unitersity of Vienna, Faculty of Psychology, Department of Psychological Assessment.
Thanks to Schuhfried.
}
\seealso{
\link{PPass}
}
