% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PPall}
\alias{PPall}
\alias{print.gpcm4pl}
\alias{summary.gpcm4pl}
\title{Estimate Person Parameters}
\usage{
PPall(respm, thres, slopes, lowerA = NULL, upperA = NULL,
  theta_start = NULL, mu = NULL, sigma2 = NULL, type = "wle",
  maxsteps = 40, exac = 0.001, ctrl = list())

\method{print}{gpcm4pl}(x, ...)

\method{summary}{gpcm4pl}(object, ...)
}
\arguments{
\item{respm}{An integer matrix, which contains the examinees reponses. An Persons x items matrix is expected.}

\item{thres}{An numeric matrix which contains the threshold parameter for each item. Currently, the first row must contain zeroes - this will change in anytime soon.}

\item{slopes}{A numeric vector, which contains the slope parameters for each item - one parameter per item is expected.}

\item{lowerA}{A numeric vector, which contains the lower asymptote parameters (kind of guessing parameter) for each item. This is only possible, in case of dichotomous responses (1,2,3,4PL model).}

\item{upperA}{numeric vector, which contains the upper asymptote parameters for each item. This is only possible, in case of dichotomous responses (1,2,3,4PL model).}

\item{theta_start}{A vector which contains a starting value for each person. Currently this is necessary to supply, but soon it will be set automatically if nothing is committed.}

\item{mu}{A numeric vector of location parameters for each person in case of MAP estimation. If nothing is submitted this is set to 0 for each person for map estimation.}

\item{sigma2}{A numeric vector of variance parameters for each person in case of MAP estimation. If nothing is submitted this is set to 1 for each person for map estimation.}

\item{type}{There are three valid entries possible: "mle", "wle" or "map". "wle" is recommanded. For a deeper understanding the papers mentioned below would be helpful for sure.}

\item{maxsteps}{The maximum number of steps the NR Algorithm will take.}

\item{exac}{How accurate are the estimates supposed to be? Default is 0.001.}

\item{ctrl}{more controls}

\item{x}{an object of class gpcm4pl which is the result of using the \code{PPall()} function}

\item{...}{just some points.}

\item{object}{an object of class gpcm4pl which is the result of using the \code{PPall()} function}
}
\description{
Compute Person Parameters for the 1,2,3,4-PL Model and for the GPCM. Choose between MLE, WLE or MAP estimation. EAP estimation will follow in the future, and drawing plausible values will also be possible soon.
}
\details{
There are not Details up to now.
}
\examples{



### data creation ##########

set.seed(1522)
# intercepts
diffpar <- seq(-3,3,length=12)
# slope parameters
sl     <- round(runif(12,0.5,1.5),2)
la     <- round(runif(12,0,0.25),2)
ua     <- round(runif(12,0.8,1),2)

# antwortmatrix (für neue MLE routine)
awm <- matrix(sample(0:1,10*12,replace=TRUE),ncol=12)



## 1PL model ##### 

# MLE estimation
res1plmle <- PPall(respm = awm,thres = diffpar, slopes = rep(1,length(diffpar)),type = "mle")
# WLE estimation
res1plwle <- PPall(respm = awm,thres = diffpar, slopes = rep(1,length(diffpar)),type = "wle")
# MAP estimation
res1plmap <- PPall(respm = awm,thres = diffpar, slopes = rep(1,length(diffpar)),type = "map")

summary(res1plmle)
summary(res1plwle)
summary(res1plmap)

## 2PL model ##### 

# MLE estimation
res2plmle <- PPall(respm = awm,thres = diffpar, slopes = sl,type = "mle")
# WLE estimation
res2plwle <- PPall(respm = awm,thres = diffpar, slopes = sl,type = "wle")
# MAP estimation
res2plmap <- PPall(respm = awm,thres = diffpar, slopes = sl,type = "map")



## 3PL model ##### 

# MLE estimation
res3plmle <- PPall(respm = awm,thres = diffpar, slopes = sl,lowerA = la,type = "mle")
# WLE estimation
res3plwle <- PPall(respm = awm,thres = diffpar, slopes = sl,lowerA = la,type = "wle")
# MAP estimation
res3plmap <- PPall(respm = awm,thres = diffpar, slopes = sl,lowerA = la,type = "map")



## 4PL model ##### 

# MLE estimation
res4plmle <- PPall(respm = awm,thres = diffpar, slopes = sl,lowerA = la,upperA=ua,type = "mle")
# WLE estimation
res4plwle <- PPall(respm = awm,thres = diffpar, slopes = sl,lowerA = la,upperA=ua,type = "wle")
# MAP estimation
res4plmap <- PPall(respm = awm,thres = diffpar, slopes = sl,lowerA = la,upperA=ua,type = "map")



################# GPCM ###########################################################################


# some threshold parameters
THRES  <- matrix(c(-2,-1.23,1.11,3.48,1
                   ,2,-1,-0.2,0.5,1.3,-0.8,1.5),nrow=2)
# slopes
sl     <- c(0.5,1,1.5,1.1,1,0.98)
awmatrix <- matrix(c(1,0,2,0,1,1,1,0,0,1
                     ,2,0,0,0,0,0,0,0,0,1,1,2,2,1,1,1,1,0,0,1),byrow=TRUE,nrow=5)


## GPCM model ##### 

# MLE estimation
resgpcmlmle <- PPall(respm = awmatrix,thres = THRES, slopes = sl,type = "mle")
# WLE estimation
resgpcmwle <- PPall(respm = awmatrix,thres = THRES, slopes = sl,type = "wle")
# MAP estimation
resgpcmmap <- PPall(respm = awmatrix,thres = THRES, slopes = sl,type = "map")



## PCM model ##### 

# MLE estimation
respcmlmle <- PPall(respm = awmatrix,thres = THRES, slopes = rep(1,ncol(THRES)),type = "mle")
# WLE estimation
respcmwle <- PPall(respm = awmatrix,thres = THRES, slopes = rep(1,ncol(THRES)),type = "wle")
# MAP estimation
respcmmap <- PPall(respm = awmatrix,thres = THRES, slopes = rep(1,ncol(THRES)),type = "map")



################# GPCM  ###########################################################################
## with different number of categories

THRES  <- matrix(c(-2,-1.23,1.11,3.48,1,2,-1,-0.2,0.5,1.3,-0.8,1.5),nrow=2)
THRES1 <- rbind(THRES,c(NA,NA,NA,NA,1.7,1))

awmatrix1 <- matrix(c(1,0,2,0,1,3,1,0,0,1,3,1,0,0
                      ,0,0,0,0,0,1,1,2,2,1,1,1,1,0,0,1),byrow=TRUE,nrow=5)

# MLE estimation
respcmlmle1 <- PPall(respm = awmatrix1,thres = THRES1, slopes = sl,type = "mle")
# WLE estimation
respcmwle1 <- PPall(respm = awmatrix1,thres = THRES1, slopes = sl,type = "wle")
# MAP estimation
respcmmap1 <- PPall(respm = awmatrix1,thres = THRES1, slopes = sl,type = "map")


################# GPCM  and 4PL mixed #############################################################
## 4PL with GPCM

THRESx <- THRES
THRESx[2,1:3] <- NA

# for the 4PL item the estimated parameters are submitted, 
# for the GPCM items the lower asymptote = 0 
# and the upper asymptote = 1.
la     <- c(0.02,0.1,0,0,0,0)
ua     <- c(0.97,0.91,1,1,1,1)


# MLE estimation
respcmlmle1 <- PPall(respm = awmatrix1,thres = THRESx, 
                     slopes = sl,lowerA = la, upperA=ua,type = "mle")
# WLE estimation
respcmwle1 <- PPall(respm = awmatrix1,thres = THRESx, 
                    slopes = sl,lowerA = la, upperA=ua,type = "wle")
# MAP estimation
respcmmap1 <- PPall(respm = awmatrix1,thres = THRESx, 
                    slopes = sl,lowerA = la, upperA=ua, type = "map")
}
\author{
Manuel Reif
}
\references{
Baker, Frank B., and Kim, Seock-Ho (2004). Item Response Theory - Parameter Estimation Techniques. CRC-Press.

Magis, D. (2013). A note on the item information function of the four-parameter logistic model. Applied Psychological Measurement, 37(4), 304-315.

Muraki, Eiji (1992). A Generalized Partial Credit Model: Application of an EM Algorithm. Applied Psychological Measurement, 16, 159-176.

Muraki, Eiji (1993). Information Functions of the Generalized Partial Credit Model. Applied Psychological Measurement, 17, 351-363.

Samejima, Fumiko (1993). The bias function of the maximum likelihood estimate of ability for the dichotomous response level. Psychometrika,  58, 195-209.

Samejima, Fumiko (1993). An approximation of the bias function of the maximum likelihood estimate of a latent variable for the general case where the item responses are discrete. Psychometrika,  58, 119-138.

Wang, S. and Wang, T. (2001). Precision of Warm's Weighted Likelihood Estimates for a Polytomous Model in Computerized Adaptive Testing. Applied Psychological Measurement, 25, 317-331.

Warm, Thomas A. (1989). Weighted Likelihood Estimation Of Ability In Item Response Theory. Psychometrika, 54, 427-450.

Yen, Y.-C., Ho, R.-G., Liao, W.-W., Chen, L.-J., & Kuo, C.-C. (2012). An empirical evaluation of the slip correction in the four parameter logistic models with computerized adaptive testing. Applied Psychological Measurement, 36, 75-87.
}
\keyword{Parameters}
\keyword{Person}

