% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzstar_mixed.R
\name{print_personfit_summary}
\alias{print_personfit_summary}
\title{Print summary of person-fit results}
\usage{
print_personfit_summary(results, alpha = 0.05)
}
\arguments{
\item{results}{Matrix from lzstar_mixed.}

\item{alpha}{Significance level for flagging.}
}
\description{
Print summary of person-fit results
}
