% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_4pl}
\alias{sim_4pl}
\title{Simulate data for 4-pl model}
\usage{
sim_4pl(beta, alpha, lowerA, upperA, theta)
}
\arguments{
\item{beta}{An numeric vector which contains the difficulty parameters for each item.}

\item{alpha}{A numeric vector, which contains the slope parameters for each item.}

\item{lowerA}{A numeric vector, which contains the lower asymptote parameters (kind of guessing parameter) for each item.}

\item{upperA}{numeric vector, which contains the upper asymptote parameters for each item.}

\item{theta}{A numeric vector which contains person parameters.}
}
\description{
This function returns a dichotomous matrix of simulated responses under given item and person parameters
}
\details{
Not details until now, because the function is very easy to use.
}
\examples{
### simulation examples

set.seed(1700)

simdat <- sim_4pl(beta=seq(-4,4,length.out=7), alpha=rep(1,7), lowerA=rep(0,7), upperA=rep(1,7), theta=c(0.231,-1.313,1.772,1.601,1.733,-2.001,0.443,3.111,-4.156))


head(simdat)
}

