% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRd.R
\name{fourpl_df}
\alias{fourpl_df}
\title{Simulated data set}
\format{A data.frame with 60 rows and 14 columns.}
\source{
simulation
}
\description{
This dataset contains data simulated with the \code{sim_4pl()} function.
}
\seealso{
\link{sim_4pl}
}
