% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRd.R
\name{pp_amt}
\alias{pp_amt}
\title{Real data set}
\format{A list with two data.frame. The first list 'daten_amt' contains 299 columns and 710 rows. Each row contains data from one person. The second data.frame 'betas' contains the difficulty parameter (1PL).}
\source{
Division of Psychological Assessment and Applied Psychometrics, Fakulty of Psychology, University of Vienna
\itemize{
\item ID: id of person
\item AGE: age in years
\item AGEMON: age in month
\item TE_GA: self-assessment ist coded with TE, otherwise GA
\item FORM: version of the test
\item TIME1: start time
\item TIME2: end time
\item REL: reliability for each person
\item i: items
}
}
\description{
This dataset contains real data from the 'Adaptive Matrices Test' (AMT).
Assesses logical reasoning as an indicator of general intelligence. The ability to identify regularities and draw logical conclusions is a very good predictor of long-term success at work.
}
\details{
The data are provided from the Unitersity of Vienna, Faculty of Psychology, Department of Psychological Assessment.
Thanks to Schuhfried.
}
\references{
\itemize{
\item Hornke, L. F., Etzel, S., & Rettig, K. (2003). Manual Adaptive Matrices Test (AMT). \emph{Moedling: SCHUHFRIED GmbH}.
}
}
\seealso{
\link{PPass}
}
