% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPass.R
\name{PPass}
\alias{PPass}
\alias{PPass.default}
\title{Person Assessment function}
\usage{
PPass(...)

\method{PPass}{default}(respdf, items = "all", mod = c("1PL", "2PL", "3PL",
  "4PL", "PCM", "GPCM", "MIXED"), fitindices = c("lz", "lz_star"), ...)
}
\arguments{
\item{respdf}{A data.frame which contains the items, and perhaps other informations. Each row is a person related resonse patter. Each column denotes a variable.}

\item{items}{A numeric (integer) vector which indicates the positions of the items in the data.frame (\code{respdf}). If \code{items} = 'all', \bold{all columns} are treated as items.}

\item{mod}{Choose your data generating model. This argument switches between the three person parameter estimating functions \code{PP_4pl}, \code{PP_gpcm} and \code{PPall}.}

\item{fitindices}{A character vector which denotes the fit indices to compute.}

\item{\ldots}{Submit arguments to the underlying functions: \code{PP_4pl}, \code{PP_gpcm} and \code{PPall} (see documentation files).}
}
\description{
Estimate Person Paramters and estimate Person Fit in one step to gain resonse pattern assessment.
}
\author{
Manuel Reif, Jan Steinfeld
}
\seealso{
\link{PP_4pl}, \link{PP_gpcm}, \link{PPall}
}

