% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_gpcm}
\alias{sim_gpcm}
\title{Simulate data for the gpcm model}
\usage{
sim_gpcm(thres, alpha, theta)
}
\arguments{
\item{thres}{An numeric matrix which contains threshold parameters for each item. The first row must contain zeroes only!}

\item{alpha}{A numeric vector, which contains the slope parameters - one parameter per item is expected.}

\item{theta}{A numeric vector which contains person parameters.}
}
\description{
This function returns a integer matrix of simulated responses under given item and person parameters
}
\details{
_
}
\examples{
### simulation examples

set.seed(1700)

simdat <- sim_4pl(beta=seq(-4,4,length.out=7), alpha=rep(1,7), lowerA=rep(0,7), upperA=rep(1,7), theta=c(0.231,-1.313,1.772,1.601,1.733,-2.001,0.443,3.111,-4.156))


head(simdat)
}

